2.-Python Script Improved and Executed.
The originally delivered script was reviewed, and a problem of unbalanced classes was detected, which leads to a prediction bias by the majority class, several possible solutions are given, these are documented within the script. In addition, an execution problem was detected when variable names were misspelled, this was corrected based on understanding the logic of the code. Finally, a first execution is carried out without problems.